Analyzing Magazine Subscription
Like every company, a magazine company is trying to understand who the regular
subscribers of the magazine are. The company’s sales have been on the decline, and we are
trying to analyze the various parameters at hand. Furthermore, we are using logistic
regression and support vector machine methodology in this project.
The dataset has 2240 Rows and 30 columns. The dataset has majorly two categorical
variables namely Education and Marital Status. The rest of the data is either numerical or
binary. The data set uses 506.6 + KB of space and has variables in object, int, and float data
types.